requirements.txt
appdirs==1.4.3
appnope==0.1.0
bleach==2.0.0
decorator==4.0.11
diff-match-patch==20121119
dj-config-url==0.1.1
dj-database-url==0.4.1
Django==1.10.4
django-cors-headers==2.0.2
django-crontab==0.7.1
django-crontab==0.7.1
django-filter==1.0.2
django-import-export==0.5.1
https://github.com/darklow/django-suit/tarball/v2
djangorestframework==3.6.2
entrypoints==0.2.2
gunicorn==19.6.0
html5lib==0.999999999
ipykernel==4.6.1
ipython==6.0.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.0.1
jupyter-console==5.1.0
jupyter-core==4.3.0
Markdown==2.6.8
MarkupSafe==1.0
mistune==0.7.4
nbconvert==5.1.1
nbformat==4.3.0
notebook==5.0.0
numpy==1.12.1
packaging==16.8
pandocfilters==1.4.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
psycopg2==2.6.2
ptyprocess==0.5.1
Pygments==2.2.0
pyowm==2.6.1
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
pyzmq==16.0.2
qtconsole==4.3.0
simplegeneric==0.8.1
six==1.10.0
tablib==0.11.4
terminado==0.6
testpath==0.3
tornado==4.5.1
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
whitenoise==3.2
widgetsnbextension==2.0.0
